Refactoring for taskhHandler
